Add useMotionSliderHook
