+++ src/components/MotionSlider/MotionSlider.js
@@ -1,7 +1,5 @@
-import { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { AnimatePresence, motion } from 'framer-motion';
-import { wrap } from 'popmotion';
 import { MdNavigateBefore, MdNavigateNext } from 'react-icons/md';
 import { LazyLoadImage } from 'react-lazy-load-image-component';
 import 'react-lazy-load-image-component/src/effects/blur.css';
@@ -9,7 +7,7 @@ import 'react-lazy-load-image-component/src/effects/blur.css';
 import ButtonIcon from 'components/ButtonIcon/ButtonIcon';
 
 import { motionSliderVariants } from 'variants';
-import { swipePower } from 'helpers';
+import { useMotionSlider } from 'hooks/useMotionSlider';
 
 import {
   container,
@@ -22,40 +20,22 @@ import {
 
 const { containerVariants, imgVariants, textVariants } = motionSliderVariants;
 
-// can swipe
-const swipingThreshold = 100000;
-const nextSlideTime = 6000;
-
 const MotionSlider = ({ data }) => {
-  const [[page, direction], setPage] = useState([0, 0]);
-  const intervalRef = useRef(0);
-  // wrap(0, 1, 0.5); // 0.5
-  // wrap(0, 1, 1.5); // 0.5
-  const currentIndex = wrap(0, data.length, page);
+  const options = {
+    nextSlideTime: 6000,
+    swipingThreshold: 100000,
+  };
+
+  const {
+    state,
+    currentIndex,
+    handleChangeSlide,
+    handleDrag,
+  } = useMotionSlider(data, options);
 
   const { srcImg, name, quote, price } = data[currentIndex];
 
-  const handleChangeSlide = (newDirection) => {
-    setPage(([prevPage]) => [prevPage + newDirection, newDirection]);
-  };
-
-  const handleDrag = (e, { offset, velocity }) => {
-    const swipe = swipePower(offset.x, velocity.x);
-    if (swipe < -swipingThreshold) {
-      handleChangeSlide(1);
-    } else if (swipe > swipingThreshold) {
-      handleChangeSlide(-1);
-    }
-  };
-
-  useEffect(() => {
-    intervalRef.current = setInterval(
-      () => handleChangeSlide(1),
-      nextSlideTime,
-    );
-
-    return () => clearInterval(intervalRef.current);
-  }, [handleChangeSlide]);
+  const [page, direction] = state;
 
   return (
     <div className={container}>
